[QA-390] Clean stale submodules when mode=full, method=copy
This change calls `git clean` appropriately for the checked-out
repository and any submodules when checking out or switching branches.

Motivation:

Consider the case where a submodule exists on branch-a but not branch-b.
When using `mode=full` and `method=copy`, a `source` directory will
exist and it will contain the repository at branch-a. If the next build
is for branch-b, the following commands will be executed:

1. git fetch -f -t <source> branch-b --progress
2. git reset --hard <tip of branch-b> --
3. git checkout -B branch-b
4. git submodule sync
5. git submodule update --init --recursive --force --checkout

Command 2 here is expected to reset the checkout to the state <tip of
branch-b>, however it will fail to remove the submodule directory and
will issue a warning like:

```
warning: unable to rmdir 'path/to/submodule': Directory not empty
```

This leaves the source directory in an unexpected state and may cause
errors in the build.

Implementation:

This change ensures that `git clean -f -f -d -x` is called on the
checkout and that `_cleanSubmodule()` is called to clean submodules
recursively also.

Note that `-f` is required twice as:
> Git will refuse to modify untracked nested git repositories (directories
> with a .git subdirectory) unless a second -f is given.
-- man git-clean

Alternatives:

Initially, the option to `git submodule deinit --all` before resetting
the state was considered but this would require re-cloning submodules in
the more common case where they exist on both the current and previous
build.