[QA-466] Accept non-virtual builder name in builders filter
Change the filtering in `is_message_needed_by_props` to accept the
non-virtual builder name for a build.

Motivation

The [Virtual Builders
pattern](https://docs.buildbot.net/latest/manual/configuration/builders.html?highlight=virtual_builder_name#virtual-builders)
allows us to schedule multiple builds of the same source and builder
with different properties. We use this to split the testing and
packaging work over many workers.

The issue is that we would like to define a GitHubStatusPush reporter
for each of these builders, however, as the `virtual_builder_name` of the
builder is calculated at build time, we can't enumerate them during
configuration.